feat/EWM-654 SelectAccountWidget separates widgets for reuse
